backend/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
ffmpeg-python
requests
pydantic
